Accept more erroneous state during development
In some cases the composeProject might not be written to the
triplestore at the time this service is called.  This racecondition
causes he mu-har-transformation-service to crash and stop processing.
A similar case could occur when services which are not attached to a
project would be logged.  This case now does not cause the logger to
crash anymore.